Move rdf.NQuad into its own file
